requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.1
gunicorn==21.2.0
h11==0.14.0
icecream==2.1.3
idna==3.6
iniconfig==2.0.0
mysql-connector==2.2.9
packaging==23.2
pluggy==1.3.0
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.17.1
pytest==7.4.3
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.24.0.post1
websockets==12.0
